+++ app/src/main/java/org/simple/clinic/navigation/v2/fragments/BaseScreen.kt
@@ -6,37 +6,34 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.get
 import androidx.viewbinding.ViewBinding
 import com.spotify.mobius.EventSource
 import com.spotify.mobius.Init
-import com.spotify.mobius.MobiusLoop
 import com.spotify.mobius.Next.noChange
 import com.spotify.mobius.Update
-import com.spotify.mobius.android.MobiusAndroid
+import com.spotify.mobius.android.MobiusLoopViewModel
+import com.spotify.mobius.functions.Consumer
 import com.spotify.mobius.rx2.RxMobius
 import io.reactivex.Observable
 import io.reactivex.ObservableTransformer
+import io.reactivex.disposables.Disposable
+import org.simple.clinic.mobius.ViewEffectsHandler
 import org.simple.clinic.mobius.ViewRenderer
 import org.simple.clinic.mobius.eventSources
 import org.simple.clinic.mobius.first
 import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.util.unsafeLazy
 
-abstract class BaseScreen<K : ScreenKey, B : ViewBinding, M : Parcelable, E, F> : Fragment() {
+abstract class BaseScreen<K : ScreenKey, B : ViewBinding, M : Parcelable, E, F, V> : Fragment() {
 
   companion object {
     private const val KEY_MODEL = "org.simple.clinic.navigation.v2.fragments.BaseScreen.KEY_MODEL"
   }
 
-  private val loop: MobiusLoop.Builder<M, E, F> by unsafeLazy {
-    RxMobius
-        .loop(createUpdate()::update, createEffectHandler())
-        .eventSources(additionalEventSources())
-  }
-
-  private val controller: MobiusLoop.Controller<M, E> by unsafeLazy {
-    MobiusAndroid.controller(loop, defaultModel(), createInit())
-  }
+  private lateinit var viewModel: MobiusLoopViewModel<M, E, F, V>
 
   protected val screenKey by unsafeLazy { ScreenKey.key<K>(this) }
 
@@ -45,8 +42,7 @@ abstract class BaseScreen<K : ScreenKey, B : ViewBinding, M : Parcelable, E, F> 
   protected val binding: B
     get() = _binding!!
 
-  val currentModel
-    get() = controller.model
+  private lateinit var eventsDisposable: Disposable
 
   abstract fun defaultModel(): M
 
@@ -54,13 +50,15 @@ abstract class BaseScreen<K : ScreenKey, B : ViewBinding, M : Parcelable, E, F> 
 
   open fun uiRenderer(): ViewRenderer<M> = NoopViewRenderer()
 
+  open fun viewEffectHandler(): ViewEffectsHandler<V> = NoopViewEffectsHandler()
+
   open fun events(): Observable<E> = Observable.never()
 
   open fun createUpdate(): Update<M, E, F> = Update { _, _ -> noChange() }
 
   open fun createInit(): Init<M, F> = Init { model -> first(model) }
 
-  open fun createEffectHandler(): ObservableTransformer<F, E> = ObservableTransformer { Observable.never() }
+  open fun createEffectHandler(viewEffectsConsumer: Consumer<V>): ObservableTransformer<F, E> = ObservableTransformer { Observable.never() }
 
   open fun additionalEventSources(): List<EventSource<E>> = emptyList()
 
@@ -77,33 +75,46 @@ abstract class BaseScreen<K : ScreenKey, B : ViewBinding, M : Parcelable, E, F> 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
 
-    val rxBridge = RxMobiusBridge(events(), uiRenderer())
-    controller.connect(rxBridge)
+    val startModel = savedInstanceState?.getParcelable(KEY_MODEL) ?: defaultModel()
 
-    if (savedInstanceState != null) {
-      val savedModel = savedInstanceState.getParcelable<M>(KEY_MODEL)!!
-      controller.replaceModel(savedModel)
+    viewModel = ViewModelProvider(viewModelStore, object : ViewModelProvider.Factory {
+      private fun loop(viewEffectsConsumer: Consumer<V>) = RxMobius
+          .loop(createUpdate(), createEffectHandler(viewEffectsConsumer))
+          .eventSources(additionalEventSources())
+
+      @Suppress("UNCHECKED_CAST")
+      override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        return MobiusLoopViewModel.create<M, E, F, V>(
+            ::loop,
+            startModel,
+            createInit()
+        ) as T
+      }
+    }).get()
+
+    val uiRenderer = uiRenderer()
+    viewModel.models.observe(viewLifecycleOwner, uiRenderer::render)
+
+    val viewEffectHandler = viewEffectHandler()
+    viewModel.viewEffects.setObserver(
+        viewLifecycleOwner,
+        { liveViewEffect -> viewEffectHandler.handle(liveViewEffect) },
+        { pausedViewEffects -> pausedViewEffects.forEach(viewEffectHandler::handle) }
+    )
+
+    eventsDisposable = events().subscribe {
+      viewModel.dispatchEvent(it!!)
     }
   }
 
   override fun onDestroyView() {
     super.onDestroyView()
-    controller.disconnect()
+    eventsDisposable.dispose()
     _binding = null
-  }
-
-  override fun onResume() {
-    super.onResume()
-    controller.start()
-  }
-
-  override fun onPause() {
-    super.onPause()
-    controller.stop()
   }
 
   override fun onSaveInstanceState(outState: Bundle) {
     super.onSaveInstanceState(outState)
-    outState.putParcelable(KEY_MODEL, controller.model)
+    outState.putParcelable(KEY_MODEL, viewModel.model)
   }
 }