+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdScreen.kt
@@ -23,6 +23,7 @@ import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 import com.jakewharton.rxbinding3.widget.editorActionEvents
 import com.jakewharton.rxbinding3.widget.textChangeEvents
+import com.spotify.mobius.functions.Consumer
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.cast
 import io.reactivex.subjects.PublishSubject
@@ -61,7 +62,8 @@ class ScanSimpleIdScreen : BaseScreen<
     ScreenScanSimpleBinding,
     ScanSimpleIdModel,
     ScanSimpleIdEvent,
-    ScanSimpleIdEffect>(), ScanSimpleIdUi, ScanSimpleIdUiActions {
+    ScanSimpleIdEffect,
+    Unit>(), ScanSimpleIdUi, ScanSimpleIdUiActions {
 
   companion object {
     private const val RATIO_4_3_VALUE = 4.0 / 3.0
@@ -129,7 +131,7 @@ class ScanSimpleIdScreen : BaseScreen<
 
   override fun createUpdate() = ScanSimpleIdUpdate(features.isEnabled(IndiaNationalHealthID), features.isEnabled(Feature.OnlinePatientLookup))
 
-  override fun createEffectHandler() = effectHandlerFactory.create(this).build()
+  override fun createEffectHandler(viewEffectsConsumer: Consumer<Unit>) = effectHandlerFactory.create(this).build()
 
   override fun onAttach(context: Context) {
     super.onAttach(context)