+++ mobius-base/src/main/java/org/simple/clinic/mobius/ViewEffectsHandler.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.mobius
+
+interface ViewEffectsHandler<V> {
+  fun handle(viewEffect: V)
+}